backend-api/requirements.txt
flask
flask-cors
paho-mqtt
pymongo
python-dotenv
